TG-0084 Доработан нотификатор + проверка

+++ internal/notifications/notificationsStruct.go
@@ -3,7 +3,7 @@ package notifications
 type NotificationsCCStruct struct {
 	IdUsr      int
 	NameUsr    string
-	IdChat     int
+	IdChat     int64
 	IdCrypto   int
 	NameCrypto string
 	Event      string
